+++ lib/lumberjack_json_device.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "lumberjack"
-require "multi_json"
+require "json"
 
 module Lumberjack
   # This Lumberjack device logs output to another device as JSON formatted text with one document per line.
@@ -35,15 +35,32 @@ module Lumberjack
     DEFAULT_TIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%6N%z"
 
     attr_accessor :formatter
+    attr_accessor :post_processor
+    attr_writer :pretty
     attr_reader :mapping
 
-    def initialize(stream_or_device, mapping: DEFAULT_MAPPING, formatter: nil, datetime_format: nil)
+    # @param stream_or_device [IO, Lumberjack::Device] The output stream or Lumberjack device to write
+    #  the JSON formatted log entries to.
+    # @param mapping [Hash] A hash where the key is the log entry field name and the value indicates how
+    #   to map the field if it exists. If the value is `true`, the field will be mapped to the same name.
+    #   If the value is an array, it will be mapped to a nested structure that follows the array elements.
+    #   If the value is a callable object, it will be called with the value and is expected to return
+    #   a hash that will be merged into the JSON document.
+    #   If the value is `false`, the field will not be included in the JSON output.
+    # @param formatter [Lumberjack::Formatter] An optional formatter to use for formatting the log entry data.
+    # @param datetime_format [String] An optional datetime format string to use for formatting the log timestamp.
+    # @param post_processor [Proc] An optional callable object that will be called with the log entry hash
+    #   before it is written to the output stream. This can be used to modify the log entry data
+    #   before it is serialized to JSON.
+    # @param pretty [Boolean] If true, the output will be formatted as pretty JSON with indentation and newlines.
+    #   The default is false, which writes each log entry as a single line JSON document.
+    def initialize(stream_or_device, mapping: DEFAULT_MAPPING, formatter: nil, datetime_format: nil, post_processor: nil, pretty: false)
       @mutex = Mutex.new
 
       @device = if stream_or_device.is_a?(Device)
         stream_or_device
       else
-        Writer.new(stream_or_device)
+        Lumberjack::Device::Writer.new(stream_or_device)
       end
 
       self.mapping = mapping
@@ -55,12 +72,17 @@ module Lumberjack
         datetime_format = DEFAULT_TIME_FORMAT if datetime_format.nil?
       end
       add_datetime_formatter!(datetime_format) unless datetime_format.nil?
+
+      @post_processor = post_processor
+
+      @pretty = !!pretty
     end
 
     def write(entry)
-      return if entry.message.nil? || entry.message == ""
+      return if entry.empty?
+
       data = entry_as_json(entry)
-      json = MultiJson.dump(data)
+      json = @pretty ? JSON.pretty_generate(data) : JSON.generate(data)
       @device.write(json)
     end
 
@@ -71,11 +93,28 @@ module Lumberjack
     attr_reader :datetime_format
 
     # Set the datetime format for the log timestamp.
+    #
+    # @param format [String] The datetime format string to use for formatting the log timestamp.
     def datetime_format=(format)
       add_datetime_formatter!(format)
     end
 
+    # Return true if the output is written in a multi-line pretty format. The default is to write each
+    # log entry as a single line JSON document.
+    #
+    # @return [Boolean]
+    def pretty?
+      !!@pretty
+    end
+
     # Set the mapping for how to map an entry to a JSON object.
+    #
+    # @param mapping [Hash] A hash where the key is the log entry field name and the value is the JSON key.
+    #   If the value is `true`, the field will be mapped to the same name
+    #   If the value is an array, it will be mapped to a nested structure.
+    #   If the value is a callable object, it will be called with the value and should return a hash that will be merged into the JSON document.
+    #   If the value is `false`, the field will not be included in the JSON output.
+    # @return [void]
     def mapping=(mapping)
       @mutex.synchronize do
         keys = {}
@@ -93,46 +132,148 @@ module Lumberjack
         @pid_key = keys.delete(:pid)
         @message_key = keys.delete(:message)
         @tags_key = keys.delete(:tags)
-        @custom_keys = keys
+        @custom_keys = keys.map do |name, key|
+          [name.to_s.split("."), key]
+        end.to_h
         @mapping = mapping
       end
     end
 
+    # Add a field mapping to the existing mappings.
+    #
+    # @param field_mapping [Hash] A hash where the key is the log entry field name and the value is the JSON key.
+    #   If the value is `true`, the field will be mapped to the same name
+    #   If the value is an array, it will be mapped to a nested structure.
+    #   If the value is a callable object, it will be called with the value and should return a hash that will be merged into the JSON document.
+    #   If the value is `false`, the field will not be included in the JSON output.
+    # @return [void]
     def map(field_mapping)
-      new_mapping = {}
-      field_mapping.each do |key, value|
-        new_mapping[key.to_sym] = value
-      end
+      new_mapping = field_mapping.transform_keys(&:to_sym)
       self.mapping = mapping.merge(new_mapping)
     end
 
     # Convert a Lumberjack::LogEntry to a Hash using the specified field mapping.
+    #
+    # @param entry [Lumberjack::LogEntry] The log entry to convert.
+    # @return [Hash] A hash representing the log entry in JSON format.
     def entry_as_json(entry)
       data = {}
-      set_attribute(data, @time_key, entry.time) unless @time_key.nil?
-      set_attribute(data, @severity_key, entry.severity_label) unless @severity_key.nil?
-      set_attribute(data, @progname_key, entry.progname) unless @progname_key.nil?
-      set_attribute(data, @pid_key, entry.pid) unless @pid_key.nil?
-      set_attribute(data, @message_key, entry.message) unless @message_key.nil?
+      set_attribute(data, @time_key, entry.time) if @time_key
+      set_attribute(data, @severity_key, entry.severity_label) if @severity_key
+      set_attribute(data, @message_key, entry.message) if @message_key
+      set_attribute(data, @progname_key, entry.progname) if @progname_key
+      set_attribute(data, @pid_key, entry.pid) if @pid_key
 
-      tags = entry.tags
-      if @custom_keys.size > 0
-        tags = (tags.nil? ? {} : tags.dup)
+      tags = dereference_tags(entry.tags) if entry.tags
+      extracted_tags = nil
+      if @custom_keys.size > 0 && !tags&.empty?
+        extracted_tags = []
         @custom_keys.each do |name, key|
-          set_attribute(data, key, tags.delete(name.to_s))
+          set_attribute(data, key, tag_value(tags, name))
+          extracted_tags << name
+        end
+
+        extracted_tags.each do |path|
+          tags = deep_remove_tag(tags, path, entry.tags)
         end
       end
 
-      unless @tags_key.nil?
+      if @tags_key
         tags ||= {}
-        set_attribute(data, @tags_key, tags)
+        if @tags_key == "*"
+          data = tags.merge(data) unless tags.empty?
+        else
+          set_attribute(data, @tags_key, tags)
+        end
       end
 
       data = @formatter.format(data) if @formatter
+      if @post_processor
+        processed_result = @post_processor.call(data)
+        data = processed_result if processed_result.is_a?(Hash)
+      end
+
       data
     end
 
     private
+
+    def dereference_tags(tags)
+      updated_tags = nil
+      remove_tags = nil
+
+      tags.each do |original_key, value|
+        updated_tags ||= tags.dup unless original_key.is_a?(String)
+        key = original_key.to_s
+
+        dot_index = key.index(".")
+        if dot_index
+          remove_tags ||= []
+          remove_tags << original_key
+          updated_tags ||= tags.dup
+          sub_key = key[dot_index + 1..-1]
+          key = key[0, dot_index]
+          existing_vals = updated_tags[key]
+          unless existing_vals.is_a?(Hash)
+            existing_vals = {}
+            updated_tags[key] = existing_vals
+          end
+          existing_vals.merge!(dereference_tags({sub_key => value}))
+        elsif value.is_a?(Enumerable)
+          updated_tags ||= tags.dup
+          updated_tags[key] = if value.is_a?(Hash)
+            dereference_tags(value)
+          else
+            value.collect { |v| v.is_a?(Hash) ? dereference_tags(v) : v }
+          end
+        elsif updated_tags
+          updated_tags[key] = value
+        end
+      end
+
+      return tags unless updated_tags
+
+      remove_tags&.each { |key| updated_tags.delete(key) }
+      updated_tags
+    end
+
+    def tag_value(tags, name)
+      return nil if tags.nil?
+      return tags[name] unless name.is_a?(Array)
+
+      val = tags[name.first]
+      return val if name.length == 1
+      return nil unless val.is_a?(Hash)
+
+      tag_value(val, name[1, name.length])
+    end
+
+    def deep_remove_tag(tags, path, original_tags)
+      return nil if tags.nil?
+
+      dup_needed = tags.equal?(original_tags)
+      key = path.first
+      val = tags[key] if path.length > 1
+      unless val.is_a?(Hash)
+        if tags.include?(key)
+          tags = tags.dup if dup_needed
+          tags.delete(key)
+        end
+        return tags
+      end
+
+      new_val = deep_remove_tag(val, path[1, path.length], original_tags[key])
+      if new_val.empty? || !new_val.equal?(val)
+        tags = tags.dup if dup_needed
+        if new_val.empty?
+          tags.delete(key)
+        else
+          tags[key] = new_val
+        end
+      end
+
+      tags
+    end
 
     def set_attribute(data, key, value)
       return if value.nil?